enemy spawning from level_info

+++ src/game/word_match/components.rs
@@ -1,13 +1,13 @@
 use bevy::prelude::Component;
 
 #[allow(dead_code)] //TODO: remove
-#[derive(PartialEq)]
+#[derive(PartialEq, Debug)]
 pub enum WordTarget {
     Reload,
     Heal,
-    Enemy(usize),
-    Move(usize)
+    Enemy(u32), // bevy stores entity ids as u32
+    Move(u32)
 }
 
-#[derive(Component)]
+#[derive(Component, Debug)]
 pub struct Word(pub WordTarget, pub String);
